% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayLowReso.R
\name{arrayLowReso}
\alias{arrayLowReso}
\title{Reduce resolution of image}
\usage{
arrayLowReso(A, pixel, func)
}
\arguments{
\item{A}{an array}

\item{pixel}{a vector determining how much pixels of initial image one pixel
of the output image contains; such as \code{c(2, 2, 2)} for a 3D array \code{A}}

\item{func}{the operation executed to get the output image; such as \code{mean}
or \code{max}}
}
\value{
an array with dim = \code{ceiling(dim(A)/pixel)}
}
\description{
Reduce resolution of image
}
\examples{
A = array(1:27, c(3, 3, 3))
pixel = c(2, 2, 2)
B = arrayLowReso(A, pixel, mean)

}
