% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BH_Extreme.R
\name{BH.asymp}
\alias{BH.asymp}
\title{BH with p-values calculated by extreme value theory}
\usage{
BH.asymp(dat, alpha, varEst = T, D_sig = NULL, outputCP = FALSE)
}
\arguments{
\item{dat}{n x p data matrix (p features, n observations)}

\item{alpha}{FDR nominal level}

\item{varEst}{logical 0 or 1; estimate variance or not (TRUE for esitmation)}

\item{D_sig}{p x 1 vector, true variances of streams,
optional only when \code{varEst = FALSE}}

\item{outputCP}{logical parameter \code{FALSE}(default); if \code{TRUE}, the
change-point location in (0, 1) corresponding to signals will be returned.}
}
\value{
A list contains:
    \item{sig}{indices of signals}
    \item{FDP}{estiamted FDP}
    \item{cps}{change-points, optional only when \code{outputCP = TRUE}}
}
\description{
BH with p-values calculated by extreme value theory
}
\examples{
  N = 120; p = 200
  data = SLIP.scp.generator(N, p)
  BH.asymp(data$dat, 0.1)

}
