% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{bootstrap.cusum}
\alias{bootstrap.cusum}
\title{Bootstrap for CUSUM statistics}
\usage{
bootstrap.cusum(N, B = 20000)
}
\arguments{
\item{N}{the length of the stream}

\item{B}{the number of simulations}
}
\value{
B x 1 vector; empirical distribution of the CUSUM
}
\description{
Bootstrap for CUSUM statistics
}
