% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLIP_MCP_generator.R
\name{SLIP.mcp.generator}
\alias{SLIP.mcp.generator}
\title{Generate n-by-p matrix with at least one changes in each abnormal datastream}
\usage{
SLIP.mcp.generator(
  n,
  p,
  cov = "Identity",
  rho = NULL,
  dist = "Gaussian",
  ratio = 0.15,
  delta = 1,
  varrho = 0.05,
  cpn = 4,
  Sigma = NULL,
  param = NULL
)
}
\arguments{
\item{n}{the number of observations}

\item{p}{the number of datastreams}

\item{cov}{covariance type \code{c("CS", "Factor", "AR1", "Block", "Sparse",
"Identity", "Given")}}

\item{rho}{optional when using \code{cov} from \code{c("CS", "AR1", "Block")},
rho in [0, 1]}

\item{dist}{noise distribution \code{c("Gaussian", "t", "exp", "chisq")}}

\item{ratio}{abnormal streams/total streams = ratio, ratio in [0, 1]}

\item{delta}{the magnitudes of changes (delta >= 0)
lie in [delta-0.1, delta+0.1] with equally probable sign from {+, -}}

\item{varrho}{the parameter avoiding the boundary problem, [0, 0.5)}

\item{cpn}{the parameter of Poisson distribution related to the number of
change-points at each of abnormal streams}

\item{Sigma}{the covariance matrix, optional only when \code{cov=('Given')}}

\item{param}{the parameter when using \code{dist} from
\code{c("t", "chisq", "exp")}}
}
\value{
data with changes in some datastreasm
    \item{dat}{n-by-p data matrix}
    \item{index}{those datastreams containing changes}
    \item{cpnum}{a vector of the number of change-points}
    \item{mu}{a list of values of changes}
    \item{loc}{a list of locations of changes}
}
\description{
Generate n-by-p matrix with at least one changes in each abnormal datastream
}
