% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLIP_thrsh_c.R
\name{SLIP.thresh.c}
\alias{SLIP.thresh.c}
\title{SLIP with thresholding (c version)}
\usage{
SLIP.thresh.c(
  dat,
  alpha,
  C = 1.5,
  r = 3,
  covEst = T,
  estMthd = "Cholesky",
  trueCov = NULL,
  outputW = FALSE,
  outputCP = FALSE
)
}
\arguments{
\item{dat}{n x p matrix (p features, n observations)}

\item{alpha}{FDR nominal level}

\item{C}{tuning parameter for mean screening; [1.4, 1.6] is suitable;
larger \code{p}, bigger \code{C}}

\item{r}{splitting ratio, (r-1) pieces versus 1 piece}

\item{covEst}{Estimate covariance or not (logical); T for Est}

\item{estMthd}{optional estimation methods \code{c("Cholesky", "POET")}}

\item{trueCov}{the true covariance matrix; only optional when \code{covEst=F}}

\item{outputW}{logical parameter \code{FALSE}(default); if \code{TRUE}, the
W-statistics and the threshold will be returned.}

\item{outputCP}{logical parameter \code{FALSE}(default); if \code{TRUE}, the
change-point location in (0, 1) corresponding to signals will be returned.}
}
\value{
A list contains:
    \item{sig}{indices of signals}
    \item{FDP}{estiamted FDP}
    \item{W}{W-statistic, optional only when \code{outputW = TRUE}}
    \item{L}{threshold, optional only when \code{outputW = TRUE}}
    \item{cps}{change-points, optional only when \code{outputCP = TRUE}}
}
\description{
Use SLIP with mean screening to detect abnormal data streams
each of which occurs at least one change.
}
\examples{
  N = 120; p = 200
  data = SLIP.scp.generator(N, p)
  SLIP.thresh.c(data$dat, 0.1)

}
